+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+2.0
+---
+ * Allow getter with a runtime default by appending the `$Default` suffix
+ * Add Jackson Serializer plugin
+
 1.1.2
 -----
  * small follow-up bugfix regarding #16